Skipping FEM LLT decomposition (Pardiso) if null space of reduced system is unchanged

+++ model/FEM/Boundaries/AtXmax.h
@@ -65,12 +65,12 @@ namespace model
                 if(eIter->second.isBoundaryElement())
                 {
                     const std::vector<int> boundaryFaces=eIter->second.boundaryFaces();
-                    for (int f=0;f<boundaryFaces.size();++f)
+                    for (size_t f=0;f<boundaryFaces.size();++f)
                     {
                         bool isExternalBoundaryFace(true);
                         
                         std::array<const Simplex<FiniteElementType::dim,0>*, SimplexTraits<FiniteElementType::dim,FiniteElementType::dim-1>::nVertices> vertices=eIter->second.simplex.child(boundaryFaces[f]).vertices();
-                        for(int v=0;v<vertices.size();++v)
+                        for(size_t v=0;v<vertices.size();++v)
                         {
                             isExternalBoundaryFace *= atx(vertices[v]->P0);
                         }
